make sure output_queue's logic is clear

+++ include/common/entry.hpp
@@ -9,8 +9,8 @@ namespace sf {
 
 // One spike entry: timestamp + neuron_id
 struct Entry {
-    uint8_t ts;         // timestamp when spike arrives
-    uint8_t neuron_id;  // which neuron fired
+    uint8_t  ts;         // timestamp when spike arrives
+    uint32_t neuron_id;  // which neuron fired (32-bit to cover large tensors)
 };
 
 } // namespace sf
